[level 2] Title: 카펫, Time: 0.14 ms, Memory: 89.1 MB -BaekjoonHub

+++ 프로그래머스/2/42842. 카펫/카펫.java
@@ -3,7 +3,7 @@ class Solution {
         int[] answer = {0,0};
         
         for(int h=3; h <= Math.sqrt(brown+yellow); h++){
-            int w = (int) ((brown+yellow) / h);
+            int w = (brown+yellow) / h;
             if(2*w + 2*h - 4 == brown) {
                 if((w-2) * (h-2) == yellow){
                     answer[0] = w;
